refactor(ColorPicker): Move duplicate picker template to a separate component
